requirements.txt
streamlit
torch
torchvision
pillow
requests
plotly